// ui_components/src/router/types.ts
import { UserRole } from '@/store/modules/auth';

export interface RouteTitle {
  title?: string;
}

export interface RouteMenu {
  icon: JSX.Element;
  text: string;
}

export interface RouteType {
  id: string;
  path: string;
  Element: ({ title }: RouteTitle) => JSX.Element;
  menu?: RouteMenu;
  roles?: UserRole[];
}